Move `ViewCtx` to a separate module `context`
