+++ xilem_web/src/lib.rs
@@ -17,6 +17,7 @@ mod app;
 mod attribute;
 mod attribute_value;
 mod class;
+mod context;
 mod element_props;
 mod events;
 mod message;
@@ -32,10 +33,11 @@ pub mod elements;
 pub mod interfaces;
 pub mod svg;
 
-pub use app::{App, ViewCtx};
+pub use app::App;
 pub use attribute::{Attr, Attributes, ElementWithAttributes, WithAttributes};
 pub use attribute_value::{AttributeValue, IntoAttributeValue};
 pub use class::{AsClassIter, Class, Classes, ElementWithClasses, WithClasses};
+pub use context::ViewCtx;
 pub use element_props::ElementProps;
 pub use message::{DynMessage, Message};
 pub use optional_action::{Action, OptionalAction};
